`initialize.sh` to build, deploy, bind, & import
- gitignore built artifacts
- run `initialize.sh` prior to running dev server

Really we want a script that's smart enough to:

- check if the contracts are already deployed
- if so, make sure the TTL is above some threshold
- check if the code hashes have changed
- if so, update existing deployed contracts, then rebind

Furthermore, we don't really want a procedural script at all. We want a
declarative config file that:

- allows describing multiple environments
  - development/local
  - testing/local
  - staging/testnet
  - production/mainnet
- allows specifying 3rd-party, already-deployed contracts

For development & testing environments, we may also want to allow the
config file to provide a way to get our own contracts into a predictable
state.

- initialize contracts
- mint tokens
- or otherwise make various contract calls

This may work better in a separate config/script, rather than the one
that only specifies contracts?

An open question is how to depend on 3rd-party contracts in development
& testing mode.

- maybe: allow specifying network, contract, & ledger sequence number of
  dependency, and automatically spoon that contract state into the local
  ledger while running the initialization for project's own contracts.